Refactor to understand remotedb packet structure
This is a large refactoring of the remoteDB internal execution of
talking with remote database devices (CDJs and Rekordbox). We now
interpret and generate the packets that we're sending and receiving as
structured objects. This adds a nice abstraction layer that should make
things much more readable with much less magic byte blobs.

Huge huge thank you to @brunchboy to his amazing work in documenting and
implementing the pioneer remote DB protocol. As well as Austin Wright

This fixes GH-03.